requirements.txt
-e .
aws-cdk.core
aws-cdk.aws_autoscaling
aws-cdk.aws_ec2
aws-cdk.aws_ecs
aws-cdk.aws_ecs_patterns
aws-cdk.aws_ecr
aws-cdk.aws_batch